+++ shell/platform/linux/BUILD.gn
@@ -138,7 +138,7 @@ source_set("flutter_linux_sources") {
     "fl_string_codec.cc",
     "fl_task_runner.cc",
     "fl_task_runner.h",
-    "fl_text_input_plugin.cc",
+    "fl_text_input_handler.cc",
     "fl_text_input_view_delegate.cc",
     "fl_texture.cc",
     "fl_texture_gl.cc",
@@ -221,7 +221,7 @@ executable("flutter_linux_unittests") {
     "fl_standard_message_codec_test.cc",
     "fl_standard_method_codec_test.cc",
     "fl_string_codec_test.cc",
-    "fl_text_input_plugin_test.cc",
+    "fl_text_input_handler_test.cc",
     "fl_texture_gl_test.cc",
     "fl_texture_registrar_test.cc",
     "fl_value_test.cc",
@@ -240,7 +240,7 @@ executable("flutter_linux_unittests") {
     "testing/mock_renderer.cc",
     "testing/mock_settings.cc",
     "testing/mock_signal_handler.cc",
-    "testing/mock_text_input_plugin.cc",
+    "testing/mock_text_input_handler.cc",
     "testing/mock_text_input_view_delegate.cc",
     "testing/mock_texture_registrar.cc",
   ]
